requirements.txt
click==8.0.3
numpy==1.21.2
pandas==1.3.4
python-dotenv==0.19.2
scikit_learn==1.0.2
setuptools==58.0.4
torch==1.10.1
transformers==4.15.0
wandb==0.12.9
wget==3.2
